Data-Science/DeepDream/requirements.txt
tensorflow>=2.12
numpy>=1.26
pillow>=10.3
matplotlib>=3.8